+++ righttyper/generate_stubs.py
@@ -3,6 +3,8 @@ from typing import TextIO
 
 from righttyper.righttyper_types import FunctionName
 
+from typing import List, Self, Sequence
+import libcst as cst
 
 def generate_stub(
     input_file: str,
@@ -90,3 +92,125 @@ def generate_variable_stub(node: ast.Assign, stub_file: TextIO) -> None:
                 stub_file.write(f"{target.id}: {annotation}\n")
             else:
                 stub_file.write(f"{target.id}: Any\n")
+
+
+
+class PyiTransformer(cst.CSTTransformer):
+    def __init__(self: Self) -> None:
+        self._needs_any = False
+
+    def leave_FunctionDef(
+        self: Self,
+        original_node: cst.FunctionDef,
+        updated_node: cst.FunctionDef
+    ) -> cst.FunctionDef:
+        return updated_node.with_changes(
+            body=cst.SimpleStatementSuite([cst.Expr(cst.Ellipsis())]),
+            leading_lines=[]
+        )
+
+    def leave_Comment(    # type: ignore[override]
+        self: Self,
+        original_node: cst.Comment,
+        updated_node: cst.Comment
+        ) -> cst.RemovalSentinel:
+        return cst.RemoveFromParent()
+
+    def value2type(self: Self, value: cst.CSTNode) -> str:
+        # FIXME not exhaustive; this should come from RightTyper typing
+        if isinstance(value, cst.BaseString):
+            return str.__name__
+        elif isinstance(value, cst.Integer):
+            return int.__name__
+        elif isinstance(value, cst.Float):
+            return float.__name__
+        elif isinstance(value, cst.Tuple):
+            return tuple.__name__
+        elif isinstance(value, cst.BaseList):
+            return list.__name__
+        elif isinstance(value, cst.BaseDict):
+            return dict.__name__
+        elif isinstance(value, cst.BaseSet):
+            return set.__name__
+        self._needs_any = True
+        return "Any"
+
+    def handle_body(self: Self, body: Sequence[cst.CSTNode]) -> List[cst.CSTNode]:
+        result: List[cst.CSTNode] = []
+        for stmt in body:
+            if isinstance(stmt, (cst.FunctionDef, cst.ClassDef, cst.If)):
+                result.append(stmt)
+            elif (isinstance(stmt, cst.SimpleStatementLine) and
+                  isinstance(stmt.body[0], (cst.Import, cst.ImportFrom))):
+                result.append(stmt)
+            elif (isinstance(stmt, cst.SimpleStatementLine) and isinstance(stmt.body[0], cst.Assign)):
+                for target in stmt.body[0].targets:
+                    type_ann = self.value2type(stmt.body[0].value)
+
+                    result.append(cst.SimpleStatementLine(body=[
+                        cst.AnnAssign(
+                            target=target.target,
+                            annotation=cst.Annotation(cst.Name(type_ann)),
+                            value=None
+                        )
+                    ]))
+
+        return result
+
+    def leave_ClassDef(
+        self: Self,
+        original_node: cst.ClassDef,
+        updated_node: cst.ClassDef
+    ) -> cst.ClassDef:
+        return updated_node.with_changes(
+            body=updated_node.body.with_changes(
+                body=self.handle_body(updated_node.body.body)
+            ),
+            leading_lines=[]
+        )
+
+    def leave_If(
+        self: Self,
+        original_node: cst.If,
+        updated_node: cst.If
+    ) -> cst.If:
+        return updated_node.with_changes(
+            body=updated_node.body.with_changes(
+                body=self.handle_body(updated_node.body.body)
+            ),
+            leading_lines=[]
+        )
+
+    def leave_Module(
+        self: Self,
+        original_node: cst.Module,
+        updated_node: cst.Module
+    ) -> cst.Module:
+        updated_node = updated_node.with_changes(
+            body=self.handle_body(updated_node.body)
+        )
+
+        if self._needs_any:
+            imports = [
+                i for i, stmt in enumerate(updated_node.body)
+                if (isinstance(stmt, cst.SimpleStatementLine) and
+                    isinstance(stmt.body[0], (cst.Import, cst.ImportFrom)))
+            ]
+
+            # TODO could check if it's already there
+            position = imports[-1]+1 if imports else 0
+
+            updated_node = updated_node.with_changes(
+                body=(*updated_node.body[:position],
+                      cst.SimpleStatementLine([
+                          cst.ImportFrom(
+                            module=cst.Name('typing'),
+                            names=[cst.ImportAlias(cst.Name('Any'))]
+                          ),
+                      ]),
+                      *updated_node.body[position:]
+                )
+            )
+
+        return updated_node
+
